Added some weaknesses, Tracking, Resolution Chart!
I can't believe I didn't put in the Tracking skill. Now it is there.

I am even less willing to believe I didn't have a resolution chart
anywhere in the repo. "table.pdf" has the resolution chart in many
different variations. Among the iterations, the leftmost column
represents the possible values a Skill Rating can hold. The other
columns represent Black, Red, Yellow, or Green Success or Gray for
Failure. The values in the individual boxes in the table either hold a
range of values or a single value.

The variants with a  range indicate the range of values that give a
specific level of success for a specific Skill Rating. For example, with
a Skill Rating of 9, I must roll between 30 and 51 to achieve a Yellow
Success.

The variants with a single value give the maximum roll you can have and
still achieve a certain level of success for a particular Skill rating.
For example, with a Skill Rating of 11 to 15, I must roll 33 or under to
get a Red Success.

Lastly, I have added some General Weaknesses. I may fill out the rest of
that section in the Weaknesses chapter before the night is through.