Added a small section on creating skills
The referee could also add Quirks or Weaknesses, but I don't think that
deserves a note like Skills do. Some skills in ZeFRS seemed like a tax,
where you'd put in one point, and subsequent points were effectively
meaningless. I would prefer the referee to not add that to their table,
as it would make better sense as a Quirk or Weakness.