Added more to the Movement table
I am going to have to seriously reconsider the Movement Rate table in
the book. I used the numbers from ZeFRS blindly, as I thought that the
logic was sound in the movement rules themselves.

Sadly, the numbers given in the ZeFRS seem ridiculously low for the
character's movement rates. A character at maximum possible Speed would
be able to sprint at around 8 miles an hour. The fastest marathon has
been run in slightly over 2 hours. This corresponds to 13-mile-an-hour
speeds for 2 hours of time. A character at maximum possible Speed rating
should be able to sustain at least that speed for half an hour (the
amount of time they could sprint). I will be reworking the numbers
tomorrow.