Fixed the Movement Rate table
You'll notice that the characters can only walk at a similar rate to
characters in an unnamed, very famous tabletop game at the very top of
the movement rate chart here. This walking speed corresponds to a gait
of around 4 miles per hour, which really should be at the top of human
performance.

The max crawling, jogging, and sprinting speeds are based on current
records/the performance of top athletes, which determine the performance
of characters with 30 in Speed. The rest of the rates are scaled
appropriately, based on the rate at 30 Speed.

I also added a limit of 16 hours to the possible number of hours you can
jog in a day. It seemed weird that a character at 30 Toughness could jog
for "30 hours a day" but only walk for 16 hours a day.