Fixed "recovering from damage"
That section in the combat chapter needed some clarification, and it
also had a typo.

I realized in the shower that I never gave a means to determine which
wounds healed first. Healing the three critical areas first is in every
character's benefit. Also, this is somewhat intuitive, as areas with
greater blood flow tend to heal more quickly.