Cleanup
Finally added back a small blurb about creating skills.
Took away the capital 'L' on the word 'lore' that was in a Summoning
effect description. Not sure if I misguidedly expected to write some
section on 'Lore,' but I don't know what I would have expected that to
entail.
Lastly, reworded a sentence at the end of the Terror section to make it
a little clearer.