Finish up the character sheet, cleanup
Cleaned up some sections, most of which was minor changes related to
capitalization and clarification of some rules.

Tried to clean out any references to Fast Talk, Persuade, or Seduction,
as I don't like the concept of those as skills.

Got rid of the Tactics skill, because as much as I would love to have a
non-magical, Bard-style skill (and a Mental combat skill), I couldn't
really come up with anything I liked.