Added Handedness and Ambidexterity
I felt that handedness was something that could affect a character's
ability to perform actions, so I figured I would add it.

Of course, I wanted to include a quirk that could get rid of this
penalty. I also figured that the Special system could be used to get rid
of the penalty after character creation, if the need arose.